11/15/14
Added header picture (giant gear (named new_banner.png), changed
background, and adjusted nav ("a" in the css file) to span across the
length of the new banner.
The banner and the nav need some coding done to where they'll expand to
fit along with the size of the screen you're using.